add adyen event listener
