+++ payment-service/src/main/java/com/hedvig/paymentservice/services/segmentPublisher/TrustlyEventListener.kt
@@ -1,12 +1,8 @@
 package com.hedvig.paymentservice.services.segmentPublisher
 
 import com.hedvig.paymentservice.domain.payments.DirectDebitStatus
-import com.hedvig.paymentservice.domain.payments.enums.AdyenAccountStatus
-import com.hedvig.paymentservice.domain.payments.events.AdyenAccountCreatedEvent
-import com.hedvig.paymentservice.domain.payments.events.AdyenAccountUpdatedEvent
 import com.hedvig.paymentservice.domain.payments.events.DirectDebitConnectedEvent
 import com.hedvig.paymentservice.domain.payments.events.DirectDebitDisconnectedEvent
-import com.hedvig.paymentservice.graphQl.types.PayinMethodStatus
 import com.hedvig.paymentservice.query.member.entities.MemberRepository
 import com.hedvig.paymentservice.serviceIntergration.notificationService.NotificationService
 import com.hedvig.paymentservice.util.isUpdateForTheLatestTrustlyAccount
@@ -21,7 +17,7 @@ import org.springframework.stereotype.Component
 @Profile("customer.io")
 @ProcessingGroup("SegmentProcessorGroup")
 @Order(1)
-class SwedishEventListener(
+class TrustlyEventListener(
     val memberRepository: MemberRepository,
     val notificationService: NotificationService
 ) {
@@ -33,16 +29,6 @@ class SwedishEventListener(
     @EventHandler
     fun on(evt: DirectDebitDisconnectedEvent) {
         updateTraitsBasedOnDirectDebitStatus(DirectDebitStatus.DISCONNECTED, evt.memberId, evt.trustlyAccountId)
-    }
-
-    @EventHandler
-    fun on(e: AdyenAccountCreatedEvent) {
-        updateTraitsBasedOnAdyenAccountStatus(e.memberId, e.accountStatus)
-    }
-
-    @EventHandler
-    fun on(e: AdyenAccountUpdatedEvent) {
-        updateTraitsBasedOnAdyenAccountStatus(e.memberId, e.accountStatus)
     }
 
     private fun updateTraitsBasedOnDirectDebitStatus(status: DirectDebitStatus, memberId: String, trustlyAccountId: String) {
@@ -66,17 +52,8 @@ class SwedishEventListener(
         notificationService.updateCustomer(memberId, traits)
     }
 
-    private fun updateTraitsBasedOnAdyenAccountStatus(memberId: String, adyenAccountStatus: AdyenAccountStatus) {
-        val traits = when (PayinMethodStatus.fromAdyenAccountStatus(adyenAccountStatus)) {
-            PayinMethodStatus.ACTIVE -> mapOf(IS_CARD_CONNECTED to true)
-            PayinMethodStatus.PENDING, PayinMethodStatus.NEEDS_SETUP -> mapOf(IS_CARD_CONNECTED to false)
-        }
-        notificationService.updateCustomer(memberId, traits)
-    }
-
     companion object {
         const val IS_DIRECT_DEBIT_ACTIVATED: String = "is_direct_debit_activated"
-        const val IS_CARD_CONNECTED: String = "is_card_connected"
         val log = LoggerFactory.getLogger(this::class.java)!!
     }
 }
